requirements.txt
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.0.0
